% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Activity_heatmap}
\alias{Activity_heatmap}
\title{Plotting TF gene expresion & activity heatmap}
\usage{
Activity_heatmap(new_activity, eset)
}
\arguments{
\item{new_activity}{Matrix. TF activity matrix}

\item{eset}{ExpressionSet of gene expression data}
}
\value{
Heatmap plotting object
}
\description{
Plotting TF gene expresion & activity heatmap
}
