% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Filter.R
\name{allNet}
\alias{allNet}
\title{Compile a GSDB to a matrix with 2 columns}
\usage{
allNet(GSDB)
}
\arguments{
\item{GSDB}{List of list. Gene set database of interactions}
}
\value{
matrix. Matrix, each row containing regulators ("from"), targets ("to")
}
\description{
Compile a GSDB to a matrix with 2 columns
}
