% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Filter.R
\name{calculateMI}
\alias{calculateMI}
\title{Calculate mutual information}
\usage{
calculateMI(actMat = actMat, nbins = 16, method = "e")
}
\arguments{
\item{actMat}{numeric matrix.}

\item{nbins}{integer (optional). Number of bins Default 16}

\item{method}{MI calculation method: e: entropy (default) or i: infotheo}
}
\value{
numeric matrix (0-1). Matrix containing mutual information values
}
\description{
Mutual information between all pairs based on entropy package.
}
