% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Analysis.R
\name{DEG_Analysis_Micro}
\alias{DEG_Analysis_Micro}
\title{Helper Function For DEG Analysis of microArray Data (for a single comparison)}
\usage{
DEG_Analysis_Micro(eset, qval = 0.05)
}
\arguments{
\item{eset}{Processed gene expression data in the ExpressionSet format
batch & experimental conditions are provided in pData.}

\item{qval}{q-value cutoff for DEG analysis (default: 0.05)}
}
\value{
DEG result in the format of a list containing:
        table: table of DEG results.
        rank_vector: a vector of t-statistics for every gene.
        degs: a vector of gene symbols for DEGs.
}
\description{
Helper Function For DEG Analysis of microArray Data (for a single comparison)
}
