% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Down_Streams}
\alias{Down_Streams}
\title{Downstream Grouping Results}
\usage{
Down_Streams(
  gs_str,
  gs_db,
  eset,
  cellwidth = 4,
  cellheight = 4,
  showname = T,
  use_module = T
)
}
\arguments{
\item{gs_str}{Vector. A list of gene set names (TFs)}

\item{gs_db}{List of list. Gene set data base}

\item{eset}{ExpressionSet of gene expression data}

\item{cellwidth}{plot width (default 4)}

\item{cellheight}{plot height (default 4)}

\item{showname}{show colnames and rownames (default TRUE)}

\item{use_module}{use modularity of gene grouping (default TRUE)}
}
\value{
List: gene grouping scheme
}
\description{
Downstream Grouping Results
}
