% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Down_Streams_nofiltering}
\alias{Down_Streams_nofiltering}
\title{Downstream Grouping Results, without doing gene filtering}
\usage{
Down_Streams_nofiltering(
  gs_str,
  gs_db,
  eset,
  cellwidth = 4,
  cellheight = 4,
  showname = T
)
}
\arguments{
\item{gs_str:}{Vector. A list of gene set names (TFs)}

\item{gs_db:}{List of list. Gene set data base}

\item{eset:}{ExpressionSet of gene expression data}

\item{cellwidth:}{plot width (default 4)}

\item{cellheight:}{plot height (default 4)}

\item{showname:}{show colnames and rownames (default TRUE)}
}
\value{
List: gene grouping scheme
}
\description{
Downstream Grouping Results, without doing gene filtering
}
