% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Filter.R
\name{getAdjacencyMat}
\alias{getAdjacencyMat}
\title{Obtain the adjacency matrix from a matrix of tf-target relationships}
\usage{
getAdjacencyMat(tfLinks = tfLinks)
}
\arguments{
\item{tfLinks}{matrix. Matrix of tf-target relationships}
}
\value{
adjMat, matrix. adjacency matrix
}
\description{
Obtain the adjacency matrix from a matrix of tf-target relationships
}
