% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Selection.R
\name{GSEA_permut_R}
\alias{GSEA_permut_R}
\title{Compute ES scores for Gene Set Enrichment Analysis (GSEA) with a new permutation method (using the original GSEA algorithm)}
\usage{
GSEA_permut_R(sim_all, gs, stats_vector)
}
\arguments{
\item{sim_all}{a matrix of permutated gene lists}

\item{gs}{a vector of genes in the gene set}

\item{stats_vector}{a vector of DEG statistics for every gene in gene_list (rank_vector in the DEG results)
Absolute values of the t-statistics are required for the desired performance.}
}
\value{
tmp_sim_sgeas: a vector of ES values for all permutated gene lists
}
\description{
The function uses the original GSEA enrichment score calculation but using the new permutation method.
             Here, the gene symbols/names are permutated without changing the ranking vector (stats_vector).
}
