% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Selection.R
\name{GSEA_permut_R_revised}
\alias{GSEA_permut_R_revised}
\title{Compute ES scores for Gene Set Enrichment Analysis (GSEA) with a new permutation method (using a revised algorithm)}
\usage{
GSEA_permut_R_revised(sim_all, gene_set, stats_vector, N)
}
\arguments{
\item{sim_all}{a vector of genes in the expression data}

\item{gene_set}{a vector of genes in the gene set}

\item{stats_vector}{a vector of DEG statistics for every gene in gene_list (rank_vector in the DEG results);
Absolute values of the t-statistics are required for the desired performance.}

\item{N}{total number of genes (size of sim_all)}
}
\value{
ES: enrichment score
}
\description{
To improve computational efficiency, we devised a new permutation approach by swapping stats_vector. 
             Here, the gene symbols/names are permutated without changing the ranking vector (stats_vector).
             This function becomes unused in NetAct, as a much faster c++ implementation (GSEA_permute) is provided.
}
