% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hDB}
\alias{hDB}
\title{Human transcription factor target regulatory database}
\format{
A list of 875 TFs, each element is an array of gene symbols of target genes.
}
\usage{
hDB
}
\description{
This data contains the literature-derived target genes of transcription factors
for human genome. For further details, 
see the bioRxiv preprint https://doi.org/10.1101/2022.05.06.487898
}
\keyword{datasets}
