% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Hill}
\alias{Hill}
\title{Hill function for the gene weight}
\usage{
Hill(x, ind)
}
\arguments{
\item{x}{value (adj p-value)}

\item{ind}{Hill coefficient}
}
\value{
Hill function of x
}
\description{
Hill function for the gene weight
}
