% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetAct.R
\docType{package}
\name{NetAct}
\alias{NetAct}
\title{NetAct}
\description{
A computational platform for constructing core transcription-factor regulatory networks using both transcriptomics data and literature-based transcription factor-target databases.
}
