% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Selection.R
\name{nm_pval}
\alias{nm_pval}
\title{Estimation of p-value for Gene Set Enrichment Analysis (GSEA) using a normal mixture model}
\usage{
nm_pval(permutated_vector)
}
\arguments{
\item{permutated_vector}{ESs for the original gene list ([1]) and all permutated gene lists ([2:n])}
}
\value{
nm_pal: estimated p-value
}
\description{
From the enrichment scores (ESs) of the original and permutated gene lists, 
             we estimate the p-value assuming that the distribution of the ESs is Gaussian.
}
