% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{plot_network}
\alias{plot_network}
\title{Plotting gene network}
\usage{
plot_network(tf_links = tf_links)
}
\arguments{
\item{tf_links}{a data frame of networ interactions}
}
\value{
visNetwork object
}
\description{
Plotting gene network
}
