% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Analysis.R
\name{Preprocess_counts}
\alias{Preprocess_counts}
\title{RNA-seq data pre processing}
\usage{
Preprocess_counts(counts, groups, mouse = FALSE)
}
\arguments{
\item{counts}{raw count matrix}

\item{groups}{group information for experimental conditions}

\item{mouse}{use mouse genome or not (default: FALSE)}
}
\value{
x$counts: processed count matrix
}
\description{
NetAct uses edgeR to load the count data and the group information for experimental conditions, 
             It also coverts gene symbols and remove duplicates.
}
