% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{rem_data}
\alias{rem_data}
\title{Remove Non-informative genes}
\usage{
rem_data(x)
}
\arguments{
\item{x}{gene expression matrix}
}
\value{
x: gene expression matrix without containing non-informative genes
}
\description{
Remove Non-informative genes
}
