% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Selection.R
\name{Reselect_TFs}
\alias{Reselect_TFs}
\title{Reselecting TFs using gene set enrichement analysis (GSEA) using an adjusted set of parameters (work together with TF_Selection)}
\usage{
Reselect_TFs(GSEArslt, qval = 0.05, combine_TFs = TRUE, ntop = NULL)
}
\arguments{
\item{GSEArslt}{GSEA results from TF_Selection}

\item{qval}{q-value cutoff (default: 0.05)}

\item{combine_TFs}{whether combine selected TFs from multiple comparisons or not (default: TRUE)}

\item{ntop}{the number of top genes (selection by the top genes) (default: NULL, no selection by the top genes)}
}
\value{
tfs: a vector of selected TFs
}
\description{
Reselecting TFs using gene set enrichement analysis (GSEA) using an adjusted set of parameters (work together with TF_Selection)
}
