% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Analysis.R
\name{RNAseqDegs_limma}
\alias{RNAseqDegs_limma}
\title{Helper Function For DEG Analysis of RNA-seq Data using limma + Voom}
\usage{
RNAseqDegs_limma(counts, phenodata, complist, lfc, qval = 0.05)
}
\arguments{
\item{counts}{Processed gene expression count data}

\item{phenodata}{pData that provides batch & experimental conditions}

\item{complist}{a vector of multiple comparisons in the format of contrasts in limma (e.g. c("A-B", "A-C", "B-C"))}

\item{lfc}{(optional) log fold change constraints for DEGs}

\item{qval}{q-value cutoff for DEG analysis (default: 0.05)}
}
\value{
DEresult: a list of DEG results, including those for each single comparison and those for the overall comparison.
        Each DEG result is in the format of A list containing:
        table: table of DEG results.
        rank_vector: a vector of t-statistics for every gene.
        degs: a vector of gene symbols for DEGs.
        e: expression data (CPM).
        e_batch: batch corrected expression.
}
\description{
Helper Function For DEG Analysis of RNA-seq Data using limma + Voom
}
