% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{row_norm}
\alias{row_norm}
\title{Row normalization (standardization)}
\usage{
row_norm(data)
}
\arguments{
\item{data}{gene expression matrix}
}
\value{
norm_data: standardized gene expression matrix
}
\description{
Row normalization (standardization)
}
