% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Activity.R
\name{TF_Activity}
\alias{TF_Activity}
\title{Inference of TF activity}
\usage{
TF_Activity(
  tfs,
  GSDB,
  eset,
  DErslt,
  with_weight = TRUE,
  if_module = FALSE,
  ind = 1/5,
  useCorSign = TRUE
)
}
\arguments{
\item{tfs}{a vector of selected tfs}

\item{GSDB}{gene set database (a list of gene sets, each of which is comprised of a vector genes)}

\item{eset}{expression set of gene expression data or gene expression matrix}

\item{DErslt}{DEG results}

\item{with_weight}{whether weighting factors (based on DEG p-values) are used to compute TF activity (default: TRUE)}

\item{if_module}{whether the grouping scheme (activation or inhibition) depends on module detection algorithm (default: FALSE, no need to change)}

\item{ind}{Hill coefficient parameter used in the weighting factors (default: 1/5, recommend to use 0 < ind < 1/4)}

\item{useCorSign}{allow the option to use the TF gene expression correlation to flip signs (default: TRUE)}
}
\value{
a list of results: 
        all_list: grouping scheme of all TF gene sets.
        all_activity: matrix of TF activity.
}
\description{
Inference of TF activity
}
