% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Filter.R
\name{TF_Filter}
\alias{TF_Filter}
\title{Generate network}
\usage{
TF_Filter(
  actMat,
  GSDB,
  miTh = 0.4,
  maxTf = 75,
  maxInteractions = 300,
  nbins = 16,
  corMethod = "spearman",
  useCor = FALSE,
  removeSignalling = FALSE,
  DPI = FALSE,
  ...
)
}
\arguments{
\item{actMat}{numeric matrix. Matrix containing the activities}

\item{GSDB}{List of list Gene set database of interactions}

\item{miTh}{numeric. Mutual information threhold}

\item{maxTf}{integer (optional) Default 75. Maximum number of transcription 
factors in the network. If \code{removeSignalling} is \code{TRUE} 
the actual number will be less.}

\item{maxInteractions}{integer (optional) Default 300. Maximum number of
interactions in the network.}

\item{nbins}{integer (optional) Number of bins Default 16.}

\item{corMethod}{character (optional) Method to compute correlaiton. See 
\link{\code[stats]{cor}}.}

\item{useCor}{Logical (optional) Whether to use correlation instead of
mutual information to find possible interactions. Default \code{FALSE}.}

\item{removeSignalling}{logical (optional). Whether to remove the Tfs which
are not the target of any other Tfs. Default TRUE. It is not recursive and 
the generated network might still contain some signallings tfs.}

\item{DPI}{logical (optional) Default TRUE. 
Whether to apply the data processing
inequality to remove weak edges from triangles.}
}
\value{
data.frame Contains the interactions in a dataframe listing 
source tf, target tf and interaction type (1-activation, 2-inhibition).
}
\description{
Network calculated using activity and interaction database. Uses
mutual information to find possible interactions and keeps the interactions 
if they are available in the database. Sign of interaction is assigned based
on the correlation between the activities.
}
