% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{toCPM}
\alias{toCPM}
\title{convert to log10 (CPM) measurement in the RNA-Seq matrix}
\usage{
toCPM(ctMat)
}
\arguments{
\item{ctMat}{Matrix of gene expression counts}
}
\value{
mat: Matrix of CPM gene expression
}
\description{
convert to log10 (CPM) measurement in the RNA-Seq matrix
}
