% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Filter.R
\name{TF_Filter}
\alias{TF_Filter}
\title{Generate network}
\usage{
TF_Filter(
  actMat,
  GSDB,
  miTh = 0.4,
  maxTf = 75,
  maxInteractions = 300,
  nbins = 16,
  miMethod = "2d",
  corMethod = "spearman",
  useCor = FALSE,
  removeSignalling = FALSE,
  DPI = FALSE,
  nameFile = NULL,
  ...
)
}
\arguments{
\item{actMat}{numeric matrix. Matrix containing the activities}

\item{GSDB}{List of list. Gene set database of interactions}

\item{miTh}{numeric. Mutual information threshold}

\item{maxTf}{integer (optional). Default 75. Maximum number of transcription 
factors in the network. If \code{removeSignalling} is \code{TRUE} 
the actual number will be less.}

\item{maxInteractions}{integer (optional). Default 300. Maximum number of
interactions in the network.}

\item{nbins}{integer (optional). Number of bins Default 16.}

\item{miMethod}{MI calculation method: 
"2d": 2D discretization with entropy (default) 
"1d": 1D discretization with infotheo}

\item{corMethod}{character (optional). Method to compute correlation.}

\item{useCor}{Logical (optional). Whether to use correlation instead of
mutual information to find possible interactions.}

\item{removeSignalling}{logical (optional). Whether to remove the Tfs which
are not the target of any other Tfs. Default TRUE. It is not recursive and 
the generated network might still contain some signalling tfs.}

\item{DPI}{logical (optional). Default FALSE. 
Whether to apply the data processing
inequality to remove weak edges from triangles.}

\item{nameFile}{character (optional). Ouput file name. Default NULL (no file output).}

\item{...}{two additional parameters passed from applyDPI (default: miDiff = 0, minMiTh = 0.5)}
}
\value{
data.frame. Contains the interactions in a dataframe listing.
source tf, target tf and interaction type (1-activation, 2-inhibition).
}
\description{
Network calculated using activity and interaction database. Uses
mutual information to find possible interactions and keeps the interactions 
if they are available in the database. Sign of interaction is assigned based
on the correlation between the activities.
}
